refactory(todo): restructure directory example page and remove type component
